Module cymbalum::input

Input and output capacities.

Input files

This module provide readers for two types of file: * Configuration files uses the YAML format to define the System or the Simulation in an human-readable way, and without writing any code; * Structures files defines the positions and the names of particles in the System or in a specific Molecule.

Reading YAML files

The read_interactions function read interactions into a System.

Reading and writing trajectories

The Trajectory type allow to read and write trajectories on the disk. One can also read a single molecule from a file using the read_molecule function.

Structs

Trajectory
A Trajectory is a file containing one or more successives simulation steps

Enums

InteractionError
Possible causes of error when reading potential files

Traits

FromToml
Convert a TOML table to a Rust type. This is the trait to implement in order to use the input files.

FromTomlWithPairs
Convert a TOML table and a PairPotential to a Rust type. This is intended to be used by potential computation mainly.

Functions

guess_bonds
Guess the bonds in a system

read_interactions
Read interactions from the TOML file at path, and add them to the system. For a full documentation of the input files syntax, see the user manual.

read_molecule
Read a the first molecule from the file at path. If no bond information exists in the file, bonds are guessed.

Type Definitions

TrajectoryResult
Result type for all Trajectory operations