Module cymbalum::potentials

Energy and forces computation in simulations.

In order to compute an energy or a force, two things are needed: a function mapping distance/angles to energy/force, and a way to compute theses functions. The functions implements the PotentialFunction trait, and the PotentialComputation trait provides ways to effectively compute the energy: with a cutoff, from a table, etc.

Reexports

pub use self::functions::PotentialFunction;

pub use self::functions::{PairPotential, AnglePotential, DihedralPotential};

pub use self::functions::{NullPotential, LennardJones, Harmonic, CosineHarmonic};

pub use self::functions::Torsion;

pub use self::computations::Computation;

pub use self::computations::{TableComputation, CutoffComputation};

pub use self::restrictions::{PairRestriction, RestrictionInfo};

pub use self::global::{GlobalPotential, CoulombicPotential};

Modules

computations
A potential computation is a way of computing a potential given its expression (represented by a PotentialFunction). The same potential can be computed either direcly, or using a cutoff, or by a table interpolation, ...

functions
Potentials traits and implementations.

global
Global potential are potentials acting on the whole system at once

restrictions
Encoding restrictions in interactions.

Structs

Ewald
Ewald summation of the coulombic interactions. The Ewald summation is based on a separation of the coulombic potential U in two parts, using the trivial identity:

Wolf
Wolf summation for the coulombic potential, as defined in [Wolf1999]. This is a fast direct pairwise summation for coulombic potential.