Module cymbalum::potentials::functions

Potentials traits and implementations.

Structs

CosineHarmonic
Cosine harmonic potential, using the following form: $$ V(r) = \frac 12 k (\cos r - \cos x_0)2 $$ where $x_0$ is the distance equilibrium, and $k$ the elastic constant.

Harmonic
Harmonic potential, using the following form: $$ V(x) = \frac 12 k (x - x_0)2 $$ where $x_0$ is the distance equilibrium, and $k$ the elastic constant.

LennardJones
Lennard-Jones potential, using the following form: $$ V(r) = 4 \epsilon \left[ (\sigma/r){12} - (\sigma/r){6} \right] $$ where $\sigma$ is a distance constant, and $\epsilon$ an energetic constant.

NullPotential
The NullPotential always returns 0.0 as energy and force.

Torsion
Torsion potential, using the following form: $$ V(r) = k(1 + \cos(n\phi􏰈 - \delta)) $$ where $k$ is the force constant, n the periodicity of the potential, and $\delta$ the equilibrium angle.

Traits

AnglePotential
Angle potential can be expressed by implenting the AnglePotential trait.

BoxCloneAngle
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

BoxCloneDihedral
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

BoxClonePair
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

BoxClonePotential
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

DihedralPotential
Dihedral angles potential can be expressed by implenting the DihedralPotential trait.

PairPotential
Pair potential can be expressed by implenting the PairPotential trait.

PotentialFunction
A PotentialFunction is a set of two parametric functions which takes a single scalar variable and return the energy or the norm of the force corresponding to the value of that variable.