Module cymbalum::potentials::global

Global potential are potentials acting on the whole system at once

They can be coulombic potentials, or external provided potential function for example.

Structs

Ewald
Ewald summation of the coulombic interactions. The Ewald summation is based on a separation of the coulombic potential U in two parts, using the trivial identity:

Wolf
Wolf summation for the coulombic potential, as defined in [Wolf1999]. This is a fast direct pairwise summation for coulombic potential.

Traits

BoxCloneCoulombic
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

BoxCloneGlobal
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

CoulombicPotential
Electrostatic potential solver should implement the CoulombicPotential trait.

GlobalCache
Energetic cache for global potentials. This trait have an opt-in default implementation that you can use by implementing the DefaultGlobalCache trait.

GlobalPotential
The GlobalPotential trait represent a potential acting on the whole system at once.