Struct cymbalum::potentials::global::Wolf

pub struct Wolf {
    // some fields omitted
}

Wolf summation for the coulombic potential, as defined in [Wolf1999]. This is a fast direct pairwise summation for coulombic potential.

[Wolf1999]: Wolf, D. et al. J. Chem. Phys. 110, 8254 (1999).

Methods

impl Wolf

fn new(cutoff: f64) -> Wolf

Create a new Wolf summation, using a real-space cutoff of cutoff.