Trait cymbalum::potentials::global::GlobalCache

pub trait GlobalCache {
    fn move_particles_cost(&mut self, system: &System, idxes: &[usize], newpos: &[Vector3D]) -> f64;
    fn update(&mut self);
}

Energetic cache for global potentials. This trait have an opt-in default implementation that you can use by implementing the DefaultGlobalCache trait.

Example

struct MyGlobalPotential;

impl GlobalPotential for MyGlobalPotential {
   ... // Implementation
}

// Use the default cache
impl DefaultGlobalCache for MyGlobalPotential {}

Required Methods

fn move_particles_cost(&mut self, system: &System, idxes: &[usize], newpos: &[Vector3D]) -> f64

Get the cost of moving the system particles whose indexes are in idxes to newpos

fn update(&mut self)

Update the cache after a call to a *_cost function.