Module cymbalum::potentials::restrictions

Encoding restrictions in interactions.

In some force fields, some interactions are restricted to a subset of the possible pairs. For example, we may want to have Lennard-Jones interactions between atoms in different molecules only, or rescale the coulombic interactions if the particles are at 1-4 distance.

Structs

RestrictionInfo
Restriction information attached to a pair of Particles in a System

Enums

PairRestriction
Possible restrictions on the pair interactions.