Trait cymbalum::potentials::PairPotential

pub trait PairPotential: PotentialFunction + BoxClonePair {
    fn virial(&self, r: &Vector3D) -> Matrix3 { ... }
}

Potential that can be used for two body interactions, either covalent or non-covalent.

Provided Methods

fn virial(&self, r: &Vector3D) -> Matrix3

Compute the virial contribution corresponding to the distance r between the particles