Module cymbalum::simulation

Types and traits for representing simulation algorithms

Reexports

pub use self::propagator::Propagator;

pub use self::md::*;

pub use self::mc::*;

pub use self::minimization::*;

pub use self::outputs::Output;

pub use self::outputs::{TrajectoryOutput, CellOutput, EnergyOutput, PropertiesOutput};

Modules

mc
Monte-Carlo Metropolis algorithms

md
Molecular dynamics algorithms.

minimization
Energy minization algorithms

outputs
Saving properties of the simulation to some stream (file or stdout mainly)

propagator
A propagator is responsible for updating the system during a simulation

Structs

Simulation
The Simulation struct holds all the needed algorithms for running the simulation. It should be use together with a System to perform the simulation.