Struct cymbalum::simulation::md::BerendsenBarostat

pub struct BerendsenBarostat {
    // some fields omitted
}

Berendsen barostat integrator based on velocity-Verlet. This one neither reversible nor symplectic.

Methods

impl BerendsenBarostat

fn new(timestep: f64, pressure: f64) -> BerendsenBarostat

Create a new Berendsen barostat with an integration timestep of timestep, and a target pressure of pressure. The barostat timestep is 1000.

fn timestep(self, timestep: f64) -> BerendsenBarostat

Set the barostat timestep. It should be expressed in terms of the integration timestep. With a barostat timestep of 1000 and an integration timestep of 0.8 fs, the effective barostat timestep will be 800 fs.