Struct cymbalum::simulation::md::BerendsenThermostat

pub struct BerendsenThermostat {
    // some fields omitted
}

Berendsen thermostat.

The berendsen thermostat sets the simulation temperature by exponentially relaxing to a desired temperature. A more complete description of this algorithm can be found in the original article [1].

[1] H.J.C. Berendsen, et al. J. Chem Phys 81, 3684 (1984); doi: 10.1063/1.448118

Methods

impl BerendsenThermostat

fn new(temperature: f64, tau: f64) -> BerendsenThermostat

Create a new BerendsenThermostat acting at temperature T, with a timestep of tau times the integrator timestep.