Struct cymbalum::simulation::minimization::SteepestDescent

pub struct SteepestDescent {
    // some fields omitted
}

Steepest gradient descent for energy minization.

This algorithm is very rough, and will not converge in all the situations. However it is easy to use, and simple enough to be implemented quickly.

Methods

impl SteepestDescent

fn new() -> SteepestDescent

Create a GradientDescent with sensible default values for energy and force convergence criteria.

The default for force criterium is 1e-5 kJ/mol/Å^2, and 1e-5 kJ/mol/Å^2 for the energy criterium.

fn with_criteria(force: f64, energy: f64) -> SteepestDescent

Create a new GradientDescent with the force convergence criterium of force, and the energy convergence criterium of energy.

fn converged(&self) -> bool

Has the minization converged so far ?