Module cymbalum::simulation::outputs

Saving properties of the simulation to some stream (file or stdout mainly)

Structs

CellOutput
The CellOutput write all the components of a cell to a file . The columns in the file contains the following values: A B C α β γ.

EnergyOutput
The EnergyOutput write the energy of the system to a text file, organized as: PotentialEnergy KineticEnergy TotalEnergy.

PropertiesOutput
The PropertiesOutput write various physical properties of the system to a file. These properties are:

TrajectoryOutput
The TrajectoryOutput allow to write the trajectory of the system to a file, using any format supported by the Chemfiles library.

Traits

Output
The Output trait define the interface for all the quantities outputed by the simulation during the run. An Output can be a text or a binary data file, an image, a text log, …