Struct cymbalum::Logger

pub struct Logger {
    // some fields omitted
}

Logger with capacity to write to the standard output stream, the standard error stream or a file.

Methods

impl Logger

fn stdout()

Initialize the global logger to use the standard output stream.

fn stdout_at_level(level: LogLevel)

Initialize the global logger to use the standard output stream, with the maximum log level of level

fn stderr()

Initialize the global logger to use the standard error stream.

fn stderr_at_level(level: LogLevel)

Initialize the global logger to use the standard error stream, with the maximum log level of level

fn file<P: AsRef<Path>>(path: P) -> Result<(), Error>

Initialize the global logger to write to the file at path, creating the file if needed. If the file already exists, it is not removed.

fn file_at_level<P: AsRef<Path>>(path: P, level: LogLevel) -> Result<(), Error>

Initialize the global logger to write to the file at path, creating the file if needed. If the file already exists, it is not removed. The maximum log level is set to level.

fn target() -> Target

Get the target for log events

fn is_screen() -> bool

Is the log events target the screen?

fn is_file() -> bool

Is the log events target a file?