Module cymbalum::system::compute

Computing properties of a system

Structs

Forces
Compute all the forces acting on the system, and return a vector of force acting on each particles

KineticEnergy
Compute the kinetic energy of the system

PotentialEnergy
Compute the potential energy of the system

Pressure
Compute the virial pressure of the system. This pressure is given by the following formula: $$ p = \frac{N k_B T}{V} + \frac{1}{3V} \sum_i \vec f_i \cdot \vec r_i $$

PressureAtTemperature
Compute the pressure of the system from the virial equation, at the given temperature. This pressure is given by the following formula: $$ p = \frac{N k_B T}{V} + \frac{1}{3V} \sum_i \vec f_i \cdot \vec r_i $$

Stress
Compute the stress tensor of the system, defined by: $$ \sigma = \frac{1}{V} (\sum_i m_i v_i \otimes v_i + \sum_i \sum_{j > i} \vec r_{ij} \otimes \vec f_{ij}) $$

StressAtTemperature
Compute the stress tensor of the system from the virial equation, at the given temperature. The stress tensor is defined by $$ \sigma = \sigma = \frac{1}{V} (\sum_i m_i v_i \otimes v_i + \sum_i \sum_{j > i} \vec r_{ij} \otimes \vec f_{ij}) $$ but here the kinetic energy term is replaced by it average at the given temperature.

Temperature
Compute the instananeous temperature of the system

TotalEnergy
Compute the total energy of the system

Virial
Compute the virial tensor of the system, defined by $$ W = \sum_i \sum_{j > i} \vec r_{ij} \otimes \vec f_{ij} $$

Volume
Compute the volume of the system

Traits

Compute
The compute trait allow to compute properties of a system, whithout modifying this system. The Output type is the type of the computed property.