Struct cymbalum::system::EnergyCache

pub struct EnergyCache {
    // some fields omitted
}

This is a cache for energy computation.

Cache integrity is left up to the user of this structure: any function with update_ prefix must be called as needed to ensure cache consistency.

Methods

impl EnergyCache

fn new() -> EnergyCache

Create a new empty energy cache.

fn init(&mut self, system: &System)

Initialize the cache to be used with system. After a call to this function, the cache is only usable with the same system. To change the associated system, one must call this function again.

fn energy(&self) -> f64

Get the cached energy

fn update(&mut self, system: &mut System)

Update the cache after a call to a EnergyCache::*_cost function or EnergyCache::unused.

fn unused(&mut self)

This function should be called whenever the cache is not used, but one still want it to be updated. Future call to EnergyCache::update will recompute the full cache.

impl EnergyCache

fn move_particles_cost(&mut self, system: &System, idxes: Vec<usize>, newpos: &[Vector3D]) -> f64

Get the cost of moving the set of particles with indexes in idxes to newpos. This function DO NOT update the cache, the update_particles_moved function MUST be called if the particles are effectively moved.