Struct cymbalum::system::PeriodicTable

pub struct PeriodicTable;

The PeriodicTable struct give access to elements information, and give a way to register new elements in the list.

Methods

impl PeriodicTable

fn add_element(element: ElementData)

Add a new element in the elements list. The additional list is searched before the PERIODIC_TABLE array, so adding a new element here will make it overwrite the real one.

This function use a RwLock to protect itself against race conditions, and will block. As such, it should only be called by one thread, and preferably before any parallel region.

fn name<S: AsRef<str>>(name: S) -> Option<&'static str>

Get the name of an element from it's symbol.

fn mass<S: AsRef<str>>(name: S) -> Option<f32>

Get the mass of an element from it's symbol.

fn covalent<S: AsRef<str>>(name: S) -> Option<f32>

Get the covalent radius of an element from it's symbol.

fn vdw<S: AsRef<str>>(name: S) -> Option<f32>

Get the Van der Waals radius of an element from it's symbol.