Struct cymbalum::system::System

pub struct System {
    // some fields omitted
}

The System type hold all the data about a simulated system.

This data contains:

an unit cell, containing the system;

a list of particles in the system;

a list of molecules in the system;

a list of interactions, associating particles kinds and potentials

a hash map associating particles names and particles kinds.

In the implementation, the particles contained in a molecule are guaranted to be contiguous in memory. This allow for faster access when iterating over molecules, and easier molecule removal in the system.

Methods

impl System

fn new() -> System

Create a new empty System

fn from_cell(cell: UnitCell) -> System

Create an empty system with a specific UnitCell

fn step(&self) -> u64

Get the current step of the system

fn increment_step(&mut self)

Increment the system step

impl System

Topology and particles related functions

fn moltype(&self, molid: usize) -> u64

Get the type of the molecule at index molid. This type is a hash of the atoms names, and the set of bonds in the molecule. This means that two molecules will have the same type if and only if they contains the same atoms and the same bonds, in the same order.

fn molecules_with_moltype(&self, moltype: u64) -> Vec<usize>

Get a list of molecules with moltype molecule type.

fn molecule_containing(&self, i: usize) -> &Molecule

Get the molecule containing the particle i

fn are_in_same_molecule(&self, i: usize, j: usize) -> bool

Check if the particles at indexes i and j are in the same molecule

fn molecules(&self) -> &[Molecule]

Get the list of molecules in the system.

fn molecule(&self, id: usize) -> &Molecule

Get the molecule at index id

fn shortest_path(&self, i: usize, j: usize) -> u8

Get the length of the shortest bond path to go from the particle i to the particle j. This length is 0 if there is no path from i to j, 1 if i == j, 2 if there is a bond between i and j, etc.

fn remove_molecule_containing(&mut self, i: usize)

Remove the molecule containing the particle at index i

fn add_bond(&mut self, i: usize, j: usize) -> Option<Permutations>

Add a bond between the particles at indexes i and j. The particles should have been added to the system before calling this.

Warning

If the bond is between two particles which are not in the same molecule, the two molecules are merged together by deplacing particles in the particles list, and thus invalidate any previously stored index. In particular, any bond, angle, dihedral or molecule is invalidated.

If molecules where merged, this function will return Some(perms), where perms contains a list of (old_index, new_index) associations. If no molecules where merged, this function will return None.

fn remove_particle(&mut self, i: usize)

Removes particle at index i and any associated bonds, angle or dihedral

fn add_particle(&mut self, p: Particle)

Insert a particle at the end of the internal list

fn size(&self) -> usize

Get the number of particles in this system

fn iter(&self) -> Iter<Particle>

Get an iterator over the Particle in this system

fn iter_mut(&mut self) -> IterMut<Particle>

Get a mutable iterator over the Particle in this system

impl System

Potentials related functions

fn energy_evaluator(&self) -> EnergyEvaluator

Get an helper struct to evaluate the energy of this system.

fn pair_potentials(&self, i: usize, j: usize) -> &[PairInteraction]

Get the list of pair potential acting between the particles at indexes i and j.

fn bond_potentials(&self, i: usize, j: usize) -> &[Box<PairPotential>]

Get the list of bonded potential acting between the particles at indexes i and j.

fn angle_potentials(&self, i: usize, j: usize, k: usize) -> &[Box<AnglePotential>]

Get the list of angle interaction acting between the particles at indexes i, j and k.

fn dihedral_potentials(&self, i: usize, j: usize, k: usize, m: usize) -> &[Box<DihedralPotential>]

Get the list of dihedral angles interaction acting between the particles at indexes i, j, k and m.

fn coulomb_potential(&self) -> Option<&RefCell<Box<CoulombicPotential>>>

Get the current coulombic solver

fn global_potentials(&self) -> &[RefCell<Box<GlobalPotential>>]

Get all the global potentials

fn add_pair_interaction(&mut self, i: &str, j: &str, potential: Box<PairPotential>)

Add the potential pair potential between the particles with names i and j.

fn add_pair_interaction_with_restriction(&mut self, i: &str, j: &str, potential: Box<PairPotential>, restriction: PairRestriction)

Add the potential pair potential between the particles with names i and j, using the restriction restriction scheme.

fn add_bond_interaction(&mut self, i: &str, j: &str, potential: Box<PairPotential>)

Add the potential bonded potentials between the particles with names i and j

fn add_angle_interaction(&mut self, i: &str, j: &str, k: &str, potential: Box<AnglePotential>)

Add the potential angle potential between the particles with names i, j, and k

fn add_dihedral_interaction(&mut self, i: &str, j: &str, k: &str, m: &str, potential: Box<DihedralPotential>)

Add the potential dihedral angle potential between the particles with names names i, j, k, and m

fn set_coulomb_interaction(&mut self, potential: Box<CoulombicPotential>)

Set the coulombic potential to potential

fn add_global_interaction(&mut self, potential: Box<GlobalPotential>)

Add a global interaction to the system

impl System

UnitCell related functions

fn cell(&self) -> &UnitCell

Get a reference to the system unit cell

fn cell_mut(&mut self) -> &mut UnitCell

Get a mutable reference to the system unit cell

fn set_cell(&mut self, cell: UnitCell)

Set the system unit cell

fn distance(&self, i: usize, j: usize) -> f64

Get the distance between the particles at indexes i and j

fn nearest_image(&self, i: usize, j: usize) -> Vector3D

Get the vector between the nearest image of particle j with respect to particle i.

fn angle(&self, i: usize, j: usize, k: usize) -> f64

Get the angle between the particles i, j and k

fn angle_and_derivatives(&self, i: usize, j: usize, k: usize) -> (f64, Vector3D, Vector3D, Vector3D)

Get the angle and the derivatives of the angle between the particles i, j and k

fn dihedral(&self, i: usize, j: usize, k: usize, m: usize) -> f64

Get the dihedral angle between the particles i, j, k and m

fn dihedral_and_derivatives(&self, i: usize, j: usize, k: usize, m: usize) -> (f64, Vector3D, Vector3D, Vector3D, Vector3D)

Get the dihedral angle and the derivatives of the dihedral angle between the particles i, j, k and m

impl System

Functions to get pysical properties of a system.

fn kinetic_energy(&self) -> f64

Get the kinetic energy of the system.

fn potential_energy(&self) -> f64

Get the potential energy of the system.

fn total_energy(&self) -> f64

Get the total energy of the system.

fn temperature(&self) -> f64

Get the temperature of the system.

fn volume(&self) -> f64

Get the volume of the system.

fn virial(&self) -> Matrix3

Get the tensorial virial of the system.

fn pressure(&self) -> f64

Get the pressure of the system from the virial equation, at the system instananeous temperature.

fn stress(&self) -> Matrix3

Get the stress tensor of the system from the virial equation, at the system instananeous temperature.

fn pressure_at_temperature(&self, temperature: f64) -> f64

Get the pressure of the system from the virial equation, at the given temperature.

fn stress_at_temperature(&self, temperature: f64) -> Matrix3

Get the stress tensor of the system from the virial equation, at the given temperature.

fn forces(&self) -> Vec<Vector3D>

Get the forces acting on all the particles in the system