Struct cymbalum::types::Array3

pub struct Array3<T>(_);

Three dimmensional array, based on ndarray

Most of the methods are simply forwarded to ndarray, so also look the doc for this crate.

Methods

impl<T: Zero + Clone> Array3<T>

fn zeros(size: (Ix, Ix, Ix)) -> Array3<T>

Create a new Array3 of the specified size filled with the Zero::zero return value.

fn resize_if_different(&mut self, size: (Ix, Ix, Ix))

Resize the array if the current size is not size, and fill the new array with zeros.

impl<T> Array3<T>

fn shape(&self) -> (Ix, Ix, Ix)

Get the shape of the array

impl<T: Clone> Array3<T>

fn assign(&mut self, value: T)

Assign the given scalar to all entry in this array

impl<T: Default> Array3<T>

fn default(size: (Ix, Ix, Ix)) -> Array3<T>

Create a new Array3 of the specified size filled with the Default::default return value.