Module cymbalum::input

Input capacities for cymbalum

This module provide input files reader for two types of files:

Configuration files uses the TOML format to define a System or a Simulation in an human-readable way, without writing any code. This type of input is further divided in potentials input files and whole simulation input files;

Structures files defines the positions and the names of particles in a System or in a specific Molecule.

Reading configuration files

The read_interactions function read interactions into a System, and the read_config function reads a whole simulation (Simulation and System objects).

Reading and writing trajectories

The Trajectory type allow to read and write trajectories on the disk. One can also read a single molecule from a file using the read_molecule function.

Structs

Trajectory
A Trajectory is a file containing one or more successives simulation steps

TrajectoryError
Possible error when reading and wrinting to trajectories

Enums

Error
Possible causes of error when reading input files

Traits

FromToml
Convert a TOML table to a Rust type.

FromTomlWithData
Convert a TOML table and some additional data to a Rust type.

FromTomlWithPairs
Convert a TOML table and a PairPotential to a Rust type. This is intended to be used by potential computation mainly.

Functions

guess_bonds
Guess the bonds in a system

read_config
Read a whole simulation input file.

read_interactions
Read interactions from the TOML file at path, and add them to the system. For a full documentation of the input files syntax, see the user manual.

read_molecule
Read a the first molecule from the file at path. If no bond information exists in the file, bonds are guessed.

Type Definitions

Result
Custom Result type for input files