Module cymbalum::potentials

Interaction potentials for energy and forces computations

Potential functions

A potential function is defined as a parametric function giving the energy and the force acting on a specific particle. These function are represented by the PotentialFunction trait, and allow to compute the energy or force norm corresponding to a scalar variable. This scalar variable will be the distance for non-bonded pair interactions; the angle for covalent angles interactions, etc.

Additional marker traits are used to specify wheter a given potential can be used for a specific interaction type. These marker traits are:

PairPotential for bonded and non-bonded two body interactions;

AnglePotential for covalent angles interactions;

DihedralPotential for covalent dihedral angles interactions.

To be able to use a specific potential in one of these contexts, the only thing to do is to implement the corresponding trait.

#[derive(Clone)]
struct OnePotential;

// OnePotential is a potential function
impl PotentialFunction for OnePotential {
    fn energy(&self, _: f64) -> f64 {1.0}
    fn force(&self, _: f64) -> f64 {0.0}
}

// It can be used for pair and dihedral potentials, but not for angles
impl PairPotential for OnePotential {}
impl DihedralPotential for OnePotential {}

Global potentials

Global potentials are potentials that have an effect on the whole system at once. They are defined by implementing the GlobalPotential trait and giving methods for energy, forces and virial contributions.

CoulombicPotential are a specific version of global potentials, that are used to compute charges-charges interactions.

Structs

CosineHarmonic
Cosine harmonic potential.

CutoffComputation
Computation of a potential with a cutoff.

Ewald
Ewald summation of the coulombic interactions.

Harmonic
Harmonic potential.

LennardJones
Lennard-Jones potential.

NullPotential
The NullPotential always returns 0 as energy and force.

RestrictionInfo
Restriction information attached to a pair of Particles in a System

TableComputation
Computation of a potential using tabulated values.

Torsion
Torsion potential.

Wolf
Wolf summation for coulombic interactions.

Enums

PairRestriction
Possible restrictions on the pair interactions.

Traits

AnglePotential
Potential that can be used for molecular angles.

Computation
Methods for energy and forces computation.

CoulombicPotential
Electrostatic potential solver.

DihedralPotential
Potential that can be used for molecular dihedral angles.

GlobalPotential
Potential acting on the whole system at once.

PairPotential
Potential that can be used for two body interactions, either covalent or non-covalent.

PotentialFunction
A set of two parametric functions which takes a single scalar variable and return the corresponding energy or norm of the force.